fixed [RookMoveValidator] lossy conversion of int into byte (while performing arthimetic operation)

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/validators/RookMoveValidator.java
@@ -2,7 +2,7 @@ package com.thinking.machines.chess.server.validators;
 import javax.swing.*;
 public class RookMoveValidator
 {
-public static byte validateMove(int startRowIndex,int startColumnIndex,int destinationRowIndex,int destinationColumnIndex,byte[][] board)
+public static byte validateMove(byte startRowIndex,byte startColumnIndex,byte destinationRowIndex,byte destinationColumnIndex,byte[][] board)
 {
 if(startRowIndex!=destinationRowIndex && startColumnIndex!=destinationColumnIndex) return 0;//restricting rook diagonal movement
 byte tile;
@@ -10,14 +10,14 @@ if(startColumnIndex==destinationColumnIndex)//vertical movement
 {
 if(startRowIndex<destinationRowIndex)	
 {
-for(int e=startRowIndex+1;e<destinationRowIndex;e++)
+for(byte e=(byte)(startRowIndex+1);e<destinationRowIndex;e++)
 {
 tile=board[e][startColumnIndex];
 if(tile!=0) return 0;
 }
 }else
 {
-for(int e=startRowIndex-1;e>destinationRowIndex;e--)
+for(byte e=(byte)(startRowIndex-1);e>destinationRowIndex;e--)
 {
 tile=board[e][startColumnIndex];
 if(tile!=0) return 0;
@@ -27,14 +27,14 @@ if(tile!=0) return 0;
 {
 if(startColumnIndex<destinationColumnIndex)
 {
-for(int f=startColumnIndex+1;f<destinationColumnIndex;f++)
+for(byte f=(byte)(startColumnIndex+1);f<destinationColumnIndex;f++)
 {
 tile=board[startRowIndex][f];
 if(tile!=0) return 0;
 }
 }else
 {
-for(int f=startColumnIndex-1;f>destinationColumnIndex;f--)
+for(byte f=(byte)(startColumnIndex-1);f>destinationColumnIndex;f--)
 {
 tile=board[startRowIndex][f];
 if(tile!=0) return 0;
